refactor: use justfile and project structure - powered by claude code and kimi k2
